[sival] Move sysrst_ctrl helper routine to its own library
This function sets up the direct IO pins and will be used in other
sysrst_ctrl tests.
